fix: 修复 StatsOverview 类型转换编译错误

+++ backend/src/main/java/com/example/followup/service/impl/StatsServiceImpl.java
@@ -77,7 +77,7 @@ public class StatsServiceImpl implements StatsService {
 
         return new StatsOverview(
                 totalPatients != null ? totalPatients : 0,
-                monthlyCompleted != null ? monthlyCompleted : 0,
+                monthlyCompleted != null ? monthlyCompleted.intValue() : 0,
                 monthlyExpected,
                 completionRate,
                 highRiskCount != null ? highRiskCount : 0,
